# swagger_Allapi_doc.yaml
consumes:
- application/json
definitions:
  ClientError:
    properties:
      message:
        description: |-
          Message of the error
          in: string
        example: bad_request
        type: string
        x-go-name: Message
    type: object
    x-go-package: visitor-management-system/docs/open_api
...
  CompanyPayload:
    properties:
      Subscription:
        $ref: '#/definitions/SubscriptionPayload'
      address:
        description: 'in: string'
        example: banani
        type: string
        x-go-name: Address
      company_name:
        description: 'in: string'
        example: vivasoft
        type: string
        x-go-name: CompanyName
      subscriber_email:
        description: 'in: string'
        example: quddusjunior1916@gmail.com
        type: string
        x-go-name: SubscriberEmail
      subscriber_name:
        description: 'in: string'
        example: nafiul quddus
        type: string
        x-go-name: SubscriberName
    type: object
    x-go-package: visitor-management-system/docs/open_api
  Genericsuccess:
    properties:
      message:
        description: |-
          Message of the success
          in: string
        example: successfull
        type: string
        x-go-name: Message
    type: object
    x-go-package: visitor-management-system/docs/open_api
  ID:
    properties:
      id:
        format: int64
        type: integer
        x-go-name: Id
    type: object
    x-go-package: visitor-management-system/docs/open_api
...
  Password:
    properties:
      confirm_password:
        type: string
        x-go-name: ConfirmPassword
      password:
        type: string
        x-go-name: Password
    type: object
    x-go-package: visitor-management-system/types
...
  ServerError:
    properties:
      message:
        description: |-
          Message of the error
          in: string
        example: server_error
        type: string
        x-go-name: Message
    type: object
    x-go-package: visitor-management-system/docs/open_api
...
  SubscriptionPayload:
    properties:
      subscription_type:
        description: 'in: string'
        example: silver
        type: string
        x-go-name: Subscription_type
    type: object
    x-go-package: visitor-management-system/docs/open_api
  Token:
    properties:
      User_Refreshtoken:
        type: string
      User_Token:
        type: string
    type: object
    x-go-package: visitor-management-system/types
...
  UnAuthorized:
    properties:
      message:
        description: |-
          Message of the error
          in: string
        example: unAuthorized
        type: string
        x-go-name: Message
    type: object
    x-go-package: visitor-management-system/docs/open_api
  User:
    properties:
      CompanyId:
        format: int64
        type: integer
      Id:
        format: int64
        type: integer
      Password:
        type: string
      email:
        type: string
        x-go-name: Email
      name:
        type: string
        x-go-name: Name
      user_type:
        type: string
        x-go-name: UserType
    type: object
    x-go-package: visitor-management-system/model
  UserCreate:
    properties:
      email:
        type: string
        x-go-name: Email
      name:
        type: string
        x-go-name: Name
    type: object
    x-go-package: visitor-management-system/docs/open_api
  UserD:
    properties:
      email:
        type: string
        x-go-name: Email
      password:
        type: string
        x-go-name: Password
      user_type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: visitor-management-system/docs/open_api
...
host: localhost:8080
info:
  contact:
    email: quddusjunior1916@gmail.com
    name: Nafiul-Quddus
  description: the purpose of this service is to provide & store all visitors of a
    company and their visit histories
  license:
    name: None
  title: VMS API.
  version: v1.0.0
paths:
  /subscriber/:
    delete:
      consumes:
      - application/json
      description: cancel the subscription
      operationId: CancelSub
      produces:
      - application/json
      - 'SecurityDefinitions:'
      - 'AuthToken:'
      - 'type: apiKey'
      - 'name: bearer'
      - 'in: header'
      responses:
        "200":
          description: Genericsuccess
          schema:
            $ref: '#/definitions/Genericsuccess'
        "400":
          description: ClientError
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: UnAuthorized
          schema:
            $ref: '#/definitions/UnAuthorized'
        "500":
          description: ServerError
          schema:
            $ref: '#/definitions/ServerError'
      tags:
      - Subscriber
  /subscriber/change-subscription:
    patch:
      consumes:
      - application/json
      description: change the subscription
      operationId: ChangeSub
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/SubscriptionPayload'
      produces:
      - application/json
      - 'SecurityDefinitions:'
      - 'AuthToken:'
      - 'type: apiKey'
      - 'name: bearer'
      - 'in: header'
      responses:
        "200":
          description: Genericsuccess
          schema:
            $ref: '#/definitions/Genericsuccess'
        "400":
          description: ClientError
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: UnAuthorized
          schema:
            $ref: '#/definitions/UnAuthorized'
        "500":
          description: ServerError
          schema:
            $ref: '#/definitions/ServerError'
      tags:
      - Subscriber
  /subscriber/registration:
    post:
      consumes:
      - application/json
      description: Create a new subscriber
      operationId: CreateSub
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/CompanyPayload'
      produces:
      - application/json
      - 'SecurityDefinitions:'
      - 'AuthToken:'
      - 'type: apiKey'
      - 'name: bearer'
      - 'in: header'
      responses:
        "201":
          description: Genericsuccess
          schema:
            $ref: '#/definitions/Genericsuccess'
        "400":
          description: ClientError
          schema:
            $ref: '#/definitions/ClientError'
        "404":
          description: ServerError
          schema:
            $ref: '#/definitions/ServerError'
        "500":
          description: ServerError
          schema:
            $ref: '#/definitions/ServerError'
      tags:
      - Subscriber
  /user/:
    delete:
      consumes:
      - application/json
      description: delete a  user
      operationId: DeleteUser
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/ID'
      produces:
      - application/json
      - 'SecurityDefinitions:'
      - 'AuthToken:'
      - 'type: apiKey'
      - 'name: bearer'
      - 'in: header'
      responses:
        "200":
          description: Genericsuccess
          schema:
            $ref: '#/definitions/Genericsuccess'
        "400":
          description: ClientError
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: UnAuthorized
          schema:
            $ref: '#/definitions/UnAuthorized'
        "500":
          description: ServerError
          schema:
            $ref: '#/definitions/ServerError'
      tags:
      - USER
  /user/change-password:
    post:
      consumes:
      - application/json
      description: change user password
      operationId: ChangePassword
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/Password'
      produces:
      - application/json
      - 'SecurityDefinitions:'
      - 'AuthToken:'
      - 'type: apiKey'
      - 'name: bearer'
      - 'in: header'
      responses:
        "200":
          $ref: '#/responses/LoginSuccess'
        "400":
          description: ClientError
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: UnAuthorized
          schema:
            $ref: '#/definitions/UnAuthorized'
        "500":
          description: ServerError
          schema:
            $ref: '#/definitions/ServerError'
      tags:
      - USER
  /user/create:
    post:
      consumes:
      - application/json
      description: Create a new user
      operationId: CreateUser
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/UserCreate'
      produces:
      - application/json
      - 'SecurityDefinitions:'
      - 'AuthToken:'
      - 'type: apiKey'
      - 'name: bearer'
      - 'in: header'
      responses:
        "200":
          description: Genericsuccess
          schema:
            $ref: '#/definitions/Genericsuccess'
        "400":
          description: ClientError
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: UnAuthorized
          schema:
            $ref: '#/definitions/UnAuthorized'
        "500":
          description: ServerError
          schema:
            $ref: '#/definitions/ServerError'
      tags:
      - USER
  /user/get-all:
    get:
      consumes:
      - application/json
      description: details of all users
      operationId: AllUser
      produces:
      - application/json
      - 'SecurityDefinitions:'
      - 'AuthToken:'
      - 'type: apiKey'
      - 'name: bearer'
      - 'in: header'
      responses:
        "200":
          $ref: '#/responses/UserDetails'
        "400":
          description: ClientError
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: UnAuthorized
          schema:
            $ref: '#/definitions/UnAuthorized'
        "500":
          description: ServerError
          schema:
            $ref: '#/definitions/ServerError'
      tags:
      - USER
  /user/login:
    patch:
      consumes:
      - application/json
      description: user login
      operationId: LoginDetails
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/UserD'
      produces:
      - application/json
      - 'SecurityDefinitions:'
      - 'AuthToken:'
      - 'type: apiKey'
      - 'name: bearer'
      - 'in: header'
      responses:
        "200":
          $ref: '#/responses/LoginSuccess'
        "400":
          description: ClientError
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: UnAuthorized
          schema:
            $ref: '#/definitions/UnAuthorized'
        "500":
          description: ServerError
          schema:
            $ref: '#/definitions/ServerError'
      tags:
      - USER
...
produces:
- application/json
responses:
...
  LoginSuccess:
    description: response after a user login
    schema:
      $ref: '#/definitions/Token'
  UserDetails:
    description: response about all user details
    schema:
      items:
        $ref: '#/definitions/User'
      type: array
...
schemes:
- http
securityDefinitions:
  base64:
    in: header
    name: ar5go-app-key
    type: apiKey
swagger: "2.0"